requirements.txt
boto==2.49.0
cachetools==3.1.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
future==0.17.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.12.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
langdetect==1.0.7
MarkupSafe==1.1.0
numpy==1.16.0
oauthlib==3.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pybind11==2.2.4
PySocks==1.6.8
python-twitter==3.5
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
tweepy==3.7.0
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
